+++ .gitignore
@@ -3,4 +3,3 @@ target/
 src/.DS_Store
 .DS_Store
 .idea/
-output/
