+++ output/P1A_WordCount/part-r-00000
@@ -1,0 +1,15 @@
+bat	2
+cat	3
+eat	2
+fat	1
+hat	3
+jat	1
+kat	1
+lat	1
+mat	3
+oat	2
+pat	4
+rat	3
+sat	2
+wat	1
+zat	1
