+++ output/P1B_InMapperCombiningWordCount/part-r-00000
@@ -1,0 +1,52 @@
+a	2
+and	1
+art	1
+artists	1
+as	1
+beautiful	1
+best	1
+blue	1
+comes	1
+consciousness	1
+day	5
+development	1
+earth	1
+enhancing	1
+every	4
+far	1
+for	1
+from	1
+hard	1
+have	1
+however	1
+if	1
+important	1
+is	7
+it	1
+let	2
+life	2
+little	1
+live	1
+look	1
+many	1
+mars	1
+meditate	1
+meditation	1
+moon	2
+of	1
+outer	1
+pays	1
+red	1
+so	2
+space	1
+sun	2
+the	2
+there	1
+us	2
+very	2
+what	1
+white	1
+wonderful	1
+world	1
+yellow	1
+you	1
