+++ src/main/java/Main.java
@@ -1,48 +1,53 @@
-import inmapperCombiningWordcount.InMapperCombiningWordCountJob;
-import mapreduce.MapReduceJob;
+import inMapperCombiningWordCount.InMapperCombiningWordCountJob;
+import mapReduce.MapReduceJob;
 import org.apache.hadoop.util.ToolRunner;
-import wordcount.WordCountJob;
+import wordCount.WordCountJob;
 
 public class Main {
-    public static void main(String[] args) throws Exception {
-        if(args.length < 3) {
-            System.out.println("Enter command such as: bin/hadoop jar hadoopProject.jar Main <p1a,p1b,p1c,p1d,p2,p3,p4> <inputPath> <outputPath> [<numReducers>]");
-            return;
-        }
+	public static void main(String[] args) throws Exception {
+		if (args.length < 3) {
+			System.out.println("Enter command such as: bin/hadoop jar hadoopProject.jar Main <p1a,p1b,p1c,p1d,p2,p3,p4> <inputPath> <outputPath> [<numReducers>]");
+			return;
+		}
 
-        MapReduceJob<?, ?, ?, ?, ?> job = null;
-        //arguments: <p1a,p1b,p1c,p1d,p2,p3,p4> <inputPath> <outputPath> [<numReducers>]
-        String jobName = args[0];
-        switch (jobName) {
-            case "p1a":
-                job = new WordCountJob("WordCountJob");
-                break;
+		MapReduceJob<?, ?, ?, ?, ?> job = null;
+		//arguments: <p1a,p1b,p1c,p1d,p2,p3,p4> <inputPath> <outputPath> [<numReducers>]
+		String jobName = args[0];
+		switch (jobName) {
+			case "p1a":
+				job = new WordCountJob("P1A_WordCount");
+				break;
 
-            case "p1b":
-                job = new InMapperCombiningWordCountJob("InMapperCombiningWordCountJob");
-                break;
+			case "p1b":
+				job = new InMapperCombiningWordCountJob("P1B_InMapperCombiningWordCount");
+				break;
 
-            case "p1c":
-                break;
+			case "p1c":
+				break;
 
-            case "p1d":
-                break;
+			case "p1d":
+				break;
 
-            case "p2":
-                break;
+			case "p2":
+				break;
 
-            case "p3":
-                break;
+			case "p3":
+				break;
 
-            case "p4":
-                break;
+			case "p4":
+				break;
 
-            default:
-                break;
-        }
+			default:
+				break;
+		}
 
-        int exitCode = ToolRunner.run(job, args);
+		if (job == null) {
+			System.out.println("Missing target. Enter command such as: bin/hadoop jar hadoopProject.jar Main <p1a,p1b,p1c,p1d,p2,p3,p4> <inputPath> <outputPath> [<numReducers>]");
+			return;
+		}
 
-        System.exit(exitCode);
-    }
+		int exitCode = ToolRunner.run(job, args);
+
+		System.exit(exitCode);
+	}
 }
