+++ src/main/java/wordCount/WordCountReducer.java
@@ -1,4 +1,4 @@
-package wordcount;
+package wordCount;
 
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
